+++ src/to-path.ts
@@ -1,33 +1,89 @@
 import { isString, assertTruthy } from './util'
+import { parseString } from './tokenize'
+import { ITokenizeOptions } from './types'
 
-// TODO: remove the regex dance
+const parseStringOptions: ITokenizeOptions = {
+  openSymbol: '[',
+  closeSymbol: ']'
+}
+
+function isQuote(str: string): boolean {
+  return /['"`]/.test(str)
+}
+
+/**
+ * @param rawPath the raw path like ".a" or " . a"
+ */
+function normalizePath(rawPath: string) {
+  let path = rawPath.trim()
+  if (path.startsWith('.')) {
+    path = path.substr(1)
+  }
+  return path
+}
+
+function pushToRet(str: string, ret: string[]): void {
+  str = normalizePath(str)
+  if (str === '') {
+    return
+  }
+  str.split('.').forEach(s => {
+    const sTrimmed = s.trim()
+    assertTruthy(
+      sTrimmed !== '',
+      `Unexpected token. Encountered empty path when parsing ${str}`,
+      SyntaxError
+    )
+    ret.push(sTrimmed)
+  })
+}
+
+export function unquote(value: string): string {
+  const key = value.trim()
+  // in our algorithms key is always a string and never only a string of spaces
+  const firstChar = key.charAt(0)
+  const lastChar = key.substr(-1)
+  if (isQuote(firstChar) || isQuote(lastChar)) {
+    assertTruthy(
+      key.length >= 2 && firstChar === lastChar,
+      `Invalid or unexpected token ${key}`,
+      SyntaxError
+    )
+    return key.substring(1, key.length - 1)
+  }
+
+  // Normalize leading plus from numerical indices
+  if (key.charAt(0) === '+') {
+    return key.substr(1)
+  }
+
+  return key
+}
+
 export function toPath(path: string): string[] {
   assertTruthy(
     isString(path),
     `Path must be a string but it is ${typeof path}: ${path}`,
     TypeError
   )
-  if (path.trim() === '') {
+
+  path = normalizePath(path)
+  if (path === '') {
     return []
   }
 
-  const keys = path
-    // convert ["something"] to .something (considering all valid js string delimiters)
-    .replace(/\[\s*(['"`])?\s*(\S+)\s*(\1)\s*\]/g, '.$2')
-    // Strip any . from the start of the string
-    .replace(/^\./g, '')
-    .split('.')
-    .map(key => key.trim())
+  // a["b"] . c => { strings: ['a', ' . c'], values: ['"b"'] }
+  const { strings, values } = parseString(path, parseStringOptions)
 
-  // check that all pieces are syntactically correct
-  keys.forEach(key => {
-    if (key.length === 0) {
-      throw new SyntaxError(`Part of the path is empty: ${path}`)
-    }
-    if (/[\[\]'"`]/.test(key)) {
-      throw new SyntaxError(`${path} resulted to an invalid key ${key}`)
-    }
-  })
+  const ret: string[] = []
 
-  return keys
+  let i = 0
+  while (i < strings.length) {
+    pushToRet(strings[i], ret)
+    if (i < values.length) {
+      ret.push(unquote(values[i]))
+    }
+    i++
+  }
+  return ret
 }